+++ src/app/deltag/page.js
@@ -1,0 +1,5 @@
+export default function Participate() {
+  return (
+    <></>
+  );
+}
