+++ src/app/kontakt/page.js
@@ -1,0 +1,5 @@
+export default function Contact() {
+  return (
+    <></>
+  );
+}
