+++ src/app/organisationen/page.js
@@ -1,0 +1,5 @@
+export default function Organisation() {
+  return (
+    <></>
+  );
+}
