+++ src/app/projekter/page.js
@@ -1,0 +1,7 @@
+import ProjectsOverview from '@/components/ProjectsOverview';
+
+export default function ProjectsPage() {
+  return (
+      <ProjectsOverview />
+  );
+}
