--- src/components/Header.js
@@ -1,20 +1,0 @@
-'use client';
-import Link from 'next/link';
-
-export default function Header() {
-
-  return (
-    <header className={`fixed inset-0 flex h-14 items-center px-4 bg-zinc-50 z-10`}>
-      <Link href="/">
-        <svg className={`w-[142.5px] h-[24px] md:w-[190px] md:h-[32px]`} viewBox="0 0 190 32" fill="none" xmlns="http://www.w3.org/2000/svg">
-          <path className='fill-zinc-800' d="M16.4074 31.2598C12.9503 31.2598 10.0032 30.6364 7.56613 29.3895C5.1291 28.1426 3.25882 26.3715 1.95529 24.0762C0.651764 21.7809 0 19.0888 0 16C0 13.478 0.425063 11.2676 1.27519 9.36901C2.12532 7.4704 3.30133 5.88349 4.80322 4.6083C6.33344 3.33311 8.10454 2.36964 10.1165 1.71787C12.1285 1.06611 14.2963 0.740227 16.62 0.740227C18.0369 0.740227 19.4396 0.881914 20.8281 1.16529C22.245 1.42033 23.6335 1.83122 24.9937 2.39798L26.3539 2.95056L24.9937 9.284L23.251 8.98646C22.0608 8.75976 20.9415 8.58973 19.893 8.47638C18.8445 8.36303 17.8527 8.30636 16.9175 8.30636C15.2173 8.30636 13.7579 8.56139 12.5394 9.07147C11.3492 9.55321 10.4282 10.3608 9.77646 11.4943C9.12469 12.6278 8.79881 14.1581 8.79881 16.085C8.79881 17.927 9.09636 19.4288 9.69144 20.5907C10.2865 21.7242 11.1792 22.546 12.3693 23.0561C13.5595 23.5661 15.0047 23.8212 16.705 23.8212C17.4418 23.8212 18.1785 23.7928 18.9153 23.7362C19.6804 23.6511 20.4455 23.5378 21.2107 23.3961V19.7831H14.3671V13.7897H28.5643V28.0718C27.6008 28.6952 26.4531 29.2478 25.1212 29.7295C23.8177 30.183 22.415 30.5513 20.9131 30.8347C19.4112 31.1181 17.9093 31.2598 16.4074 31.2598Z" fill="#27272A" />
-          <path className='fill-zinc-800' d="M32.6914 30.8772V1.12278H56.5374V8.47638H41.2776V12.5145H54.8372V19.018H41.2776V23.4811H56.835V30.8772H32.6914Z" fill="#27272A" />
-          <path className='fill-zinc-800' d="M60.6276 30.8772V1.12278H68.2788L80.4781 16.2975H80.5631V1.12278H89.0644V30.8772H81.7958L69.2564 15.2349H69.1714V30.8772H60.6276Z" fill="#27272A" />
-          <path className='fill-zinc-800' d="M94.2923 30.8772V1.12278H102.879V11.9619H114.015V1.12278H122.602V30.8772H114.015V19.4855H102.879V30.8772H94.2923Z" fill="#27272A" />
-          <path className='fill-zinc-800' d="M130.298 31.2598L126.982 27.9868L154.569 0.740227L157.884 4.01322L130.298 31.2598ZM142.455 31.2598C139.224 31.2598 136.433 30.6364 134.081 29.3895C131.757 28.1143 129.972 26.329 128.725 24.0337C127.478 21.7384 126.855 19.0463 126.855 15.9575C126.855 12.8687 127.507 10.1908 128.81 7.9238C130.114 5.62846 131.941 3.85736 134.293 2.61051C136.645 1.36365 139.366 0.740227 142.455 0.740227C145.713 0.740227 148.505 1.37782 150.828 2.65301C153.152 3.89987 154.937 5.67096 156.184 7.9663C157.431 10.2333 158.054 12.9254 158.054 16.0425C158.054 19.103 157.403 21.7809 156.099 24.0762C154.824 26.3715 153.01 28.1426 150.658 29.3895C148.335 30.6364 145.6 31.2598 142.455 31.2598ZM142.455 23.3961C143.872 23.3961 145.09 23.1694 146.11 22.716C147.13 22.2343 147.91 21.455 148.448 20.3782C149.015 19.273 149.298 17.7994 149.298 15.9575C149.298 14.1156 149.029 12.6703 148.491 11.6219C147.981 10.545 147.215 9.77991 146.195 9.32651C145.175 8.84477 143.928 8.6039 142.455 8.6039C141.066 8.6039 139.862 8.84477 138.842 9.32651C137.821 9.77991 137.028 10.545 136.461 11.6219C135.923 12.6987 135.654 14.1722 135.654 16.0425C135.654 17.8561 135.909 19.3013 136.419 20.3782C136.957 21.455 137.736 22.2343 138.757 22.716C139.777 23.1694 141.009 23.3961 142.455 23.3961Z" fill="#27272A" />
-          <path className='fill-zinc-800' d="M162.286 30.8772V1.12278H177.461C179.699 1.12278 181.626 1.51951 183.241 2.31296C184.857 3.10641 186.104 4.22575 186.982 5.67096C187.889 7.11618 188.342 8.8306 188.342 10.8142C188.342 12.6845 187.96 14.2572 187.195 15.5324C186.458 16.7793 185.452 17.7853 184.177 18.5504C182.93 19.2872 181.499 19.8256 179.883 20.1656C178.297 20.4773 176.667 20.6332 174.995 20.6332H170.872V30.8772H162.286ZM180.521 30.8772L174.485 19.9956L182.561 18.5079L190 30.8772H180.521ZM170.872 14.6398H175.718C177.135 14.6398 178.183 14.3706 178.863 13.8322C179.543 13.2938 179.883 12.472 179.883 11.3668C179.883 10.29 179.543 9.5107 178.863 9.02896C178.212 8.51889 177.163 8.26385 175.718 8.26385H170.872V14.6398Z" fill="#27272A" />
-        </svg>
-      </Link>
-    </header>
-  );
-};
