+++ src/components/UI/Footer.js
@@ -1,0 +1,60 @@
+import Link from "next/link";
+
+const footerNavItems = [
+    { href: "/", label: "Privatlivspolitik" },
+    { href: "/", label: "Vedtægter" },
+    { href: "/", label: "Cookies" },
+];
+
+export default function Footer() {
+  return (
+    <footer className='flex flex-col md:flex-row-reverse items-center gap-6 md:gap-4 p-8'>
+
+        {/* Social Media Buttons */}
+        <div className="flex flex-1 justify-end items-center md:h-14 gap-6">
+            <a href="https://www.facebook.com/people/Genh%C3%B8r/61559553821169/" className="group w-12 h-12 rounded-full border-2 border-zinc-50 bg-zinc-50 hover:bg-zinc-800 flex items-center justify-center" target="_blank" rel="noopener noreferrer" aria-label="Besøg Genhør på Facebook">
+                <svg className="w-6 h-6 fill-zinc-800 group-hover:fill-zinc-50" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+                    <path d="M14.2 2.875C12.9734 2.875 11.797 3.36228 10.9296 4.22963C10.0623 5.09699 9.575 6.27337 9.575 7.5V10.075H7.1C6.976 10.075 6.875 10.175 6.875 10.3V13.7C6.875 13.824 6.975 13.925 7.1 13.925H9.575V20.9C9.575 21.024 9.675 21.125 9.8 21.125H13.2C13.324 21.125 13.425 21.025 13.425 20.9V13.925H15.922C16.025 13.925 16.115 13.855 16.14 13.755L16.99 10.355C16.9984 10.3218 16.9991 10.2872 16.992 10.2537C16.985 10.2202 16.9704 10.1888 16.9494 10.1618C16.9283 10.1348 16.9014 10.1129 16.8707 10.0979C16.84 10.0829 16.8062 10.075 16.772 10.075H13.425V7.5C13.425 7.39823 13.445 7.29745 13.484 7.20342C13.5229 7.10939 13.58 7.02396 13.652 6.95199C13.724 6.88003 13.8094 6.82294 13.9034 6.78399C13.9974 6.74505 14.0982 6.725 14.2 6.725H16.8C16.924 6.725 17.025 6.625 17.025 6.5V3.1C17.025 2.976 16.925 2.875 16.8 2.875H14.2Z"/>
+                </svg>
+            </a>
+            <a href="https://www.instagram.com/genhoer_dk/" className="group w-12 h-12 rounded-full border-2 border-zinc-50 bg-zinc-50 hover:bg-zinc-800 flex items-center justify-center" target="_blank" rel="noopener noreferrer" aria-label="Besøg Genhør på Instagram">
+                <svg className="w-6 h-6 fill-zinc-800 group-hover:fill-zinc-50" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+                    <path d="M13.028 2C14.153 2.003 14.724 2.009 15.217 2.023L15.411 2.03C15.635 2.038 15.856 2.048 16.123 2.06C17.187 2.11 17.913 2.278 18.55 2.525C19.21 2.779 19.766 3.123 20.322 3.678C20.8307 4.17773 21.2242 4.78247 21.475 5.45C21.722 6.087 21.89 6.813 21.94 7.878C21.952 8.144 21.962 8.365 21.97 8.59L21.976 8.784C21.991 9.276 21.997 9.847 21.999 10.972L22 11.718V13.028C22.0025 13.7574 21.9948 14.4868 21.977 15.216L21.971 15.41C21.963 15.635 21.953 15.856 21.941 16.122C21.891 17.187 21.721 17.912 21.475 18.55C21.2242 19.2175 20.8307 19.8223 20.322 20.322C19.8223 20.8307 19.2175 21.2242 18.55 21.475C17.913 21.722 17.187 21.89 16.123 21.94L15.411 21.97L15.217 21.976C14.724 21.99 14.153 21.997 13.028 21.999L12.282 22H10.973C10.2433 22.0026 9.51354 21.9949 8.78402 21.977L8.59002 21.971C8.35263 21.962 8.11529 21.9517 7.87802 21.94C6.81402 21.89 6.08802 21.722 5.45002 21.475C4.78285 21.2241 4.17846 20.8306 3.67902 20.322C3.16996 19.8224 2.77613 19.2176 2.52502 18.55C2.27802 17.913 2.11002 17.187 2.06002 16.122L2.03002 15.41L2.02502 15.216C2.00659 14.4868 1.99825 13.7574 2.00002 13.028V10.972C1.99725 10.2426 2.00459 9.5132 2.02202 8.784L2.02902 8.59C2.03702 8.365 2.04702 8.144 2.05902 7.878C2.10902 6.813 2.27702 6.088 2.52402 5.45C2.77571 4.7822 3.17024 4.17744 3.68002 3.678C4.17917 3.16955 4.78321 2.77607 5.45002 2.525C6.08802 2.278 6.81302 2.11 7.87802 2.06C8.14402 2.048 8.36602 2.038 8.59002 2.03L8.78402 2.024C9.51321 2.00623 10.2426 1.99857 10.972 2.001L13.028 2ZM12 7C10.6739 7 9.40217 7.52678 8.46449 8.46447C7.5268 9.40215 7.00002 10.6739 7.00002 12C7.00002 13.3261 7.5268 14.5979 8.46449 15.5355C9.40217 16.4732 10.6739 17 12 17C13.3261 17 14.5979 16.4732 15.5356 15.5355C16.4732 14.5979 17 13.3261 17 12C17 10.6739 16.4732 9.40215 15.5356 8.46447C14.5979 7.52678 13.3261 7 12 7ZM12 9C12.394 8.99993 12.7841 9.07747 13.1481 9.22817C13.5121 9.37887 13.8429 9.5998 14.1215 9.87833C14.4001 10.1569 14.6211 10.4875 14.772 10.8515C14.9228 11.2154 15.0005 11.6055 15.0005 11.9995C15.0006 12.3935 14.9231 12.7836 14.7723 13.1476C14.6216 13.5116 14.4007 13.8423 14.1222 14.121C13.8437 14.3996 13.513 14.6206 13.149 14.7714C12.7851 14.9223 12.395 14.9999 12.001 15C11.2054 15 10.4423 14.6839 9.8797 14.1213C9.31709 13.5587 9.00102 12.7956 9.00102 12C9.00102 11.2044 9.31709 10.4413 9.8797 9.87868C10.4423 9.31607 11.2054 9 12.001 9M17.251 5.5C16.9195 5.5 16.6016 5.6317 16.3671 5.86612C16.1327 6.10054 16.001 6.41848 16.001 6.75C16.001 7.08152 16.1327 7.39946 16.3671 7.63388C16.6016 7.8683 16.9195 8 17.251 8C17.5825 8 17.9005 7.8683 18.1349 7.63388C18.3693 7.39946 18.501 7.08152 18.501 6.75C18.501 6.41848 18.3693 6.10054 18.1349 5.86612C17.9005 5.6317 17.5825 5.5 17.251 5.5Z"/>
+                </svg>
+            </a>
+            <a href="https://linkedin.com/company/genhoer" className="group w-12 h-12 rounded-full border-2 border-zinc-50 bg-zinc-50 hover:bg-zinc-800 flex items-center justify-center" target="_blank" rel="noopener noreferrer" aria-label="Besøg Genhør på LinkedIn">
+                <svg className="w-6 h-6 fill-zinc-800 group-hover:fill-zinc-50" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+                    <path d="M6.94 5.00002C6.93974 5.53046 6.72877 6.03906 6.35351 6.41394C5.97825 6.78883 5.46944 6.99929 4.939 6.99902C4.40857 6.99876 3.89997 6.78779 3.52508 6.41253C3.1502 6.03727 2.93974 5.52846 2.94 4.99802C2.94027 4.46759 3.15124 3.95899 3.5265 3.5841C3.90176 3.20922 4.41057 2.99876 4.941 2.99902C5.47144 2.99929 5.98004 3.21026 6.35492 3.58552C6.72981 3.96078 6.94027 4.46959 6.94 5.00002ZM7 8.48002H3V21H7V8.48002ZM13.32 8.48002H9.34V21H13.28V14.43C13.28 10.77 18.05 10.43 18.05 14.43V21H22V13.07C22 6.90002 14.94 7.13002 13.28 10.16L13.32 8.48002Z"/>
+                </svg>
+            </a>
+        </div>
+
+        {/* Footer Logo */}
+        <svg className='w-[121px] h-[56px]' viewBox="0 0 121 56" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path className='fill-zinc-50' d="M10.5237 56C8.32449 56 6.44974 55.6034 4.89947 54.8103C3.34919 54.0171 2.15945 52.8904 1.33023 51.4303C0.501015 49.9702 0.086407 48.2576 0.086407 46.2928C0.086407 44.6884 0.356804 43.2824 0.897597 42.0746C1.43839 40.8668 2.18649 39.8573 3.14189 39.0461C4.11532 38.2349 5.24197 37.622 6.52185 37.2074C7.80173 36.7928 9.18075 36.5855 10.6589 36.5855C11.5602 36.5855 12.4526 36.6757 13.3358 36.8559C14.2372 37.0182 15.1205 37.2795 15.9857 37.6401L16.851 37.9916L15.9857 42.0205L14.8771 41.8312C14.12 41.687 13.408 41.5789 12.741 41.5067C12.074 41.4346 11.4431 41.3986 10.8482 41.3986C9.76661 41.3986 8.83825 41.5608 8.06311 41.8853C7.306 42.1917 6.72014 42.7055 6.30553 43.4266C5.89092 44.1476 5.68362 45.121 5.68362 46.3468C5.68362 47.5186 5.8729 48.474 6.25145 49.2131C6.63001 49.9341 7.19784 50.4569 7.95495 50.7814C8.71206 51.1058 9.63141 51.2681 10.713 51.2681C11.1817 51.2681 11.6504 51.25 12.1191 51.214C12.6058 51.1599 13.0925 51.0878 13.5792 50.9977V48.6993H9.22582V44.8867H18.2571V53.972C17.6442 54.3686 16.9141 54.7201 16.0669 55.0266C15.2376 55.315 14.3453 55.5493 13.3899 55.7296C12.4345 55.9099 11.4791 56 10.5237 56Z" fill="#FAFAFA"/>
+            <path className='fill-zinc-50' d="M20.8825 55.7567V36.8289H36.0517V41.5067H26.3445V44.0755H34.9701V48.2126H26.3445V51.0517H36.241V55.7567H20.8825Z" fill="#FAFAFA"/>
+            <path className='fill-zinc-50' d="M38.6536 55.7567V36.8289H43.5208L51.2812 46.482H51.3353V36.8289H56.7432V55.7567H52.1194L44.1427 45.806H44.0886V55.7567H38.6536Z" fill="#FAFAFA"/>
+            <path className='fill-zinc-50' d="M60.0689 55.7567V36.8289H65.5309V43.724H72.6153V36.8289H78.0773V55.7567H72.6153V48.51H65.5309V55.7567H60.0689Z" fill="#FAFAFA"/>
+            <path className='fill-zinc-50' d="M82.9731 56L80.8641 53.918L98.4128 36.5855L100.522 38.6676L82.9731 56ZM90.7065 56C88.6515 56 86.8759 55.6034 85.3797 54.8103C83.9015 53.9991 82.7658 52.8634 81.9727 51.4033C81.1795 49.9431 80.7829 48.2306 80.7829 46.2657C80.7829 44.3008 81.1975 42.5973 82.0268 41.1552C82.856 39.6951 84.0187 38.5684 85.5149 37.7753C87.0111 36.9821 88.7416 36.5855 90.7065 36.5855C92.7795 36.5855 94.5551 36.9911 96.0333 37.8023C97.5115 38.5955 98.6471 39.7221 99.4403 41.1823C100.233 42.6244 100.63 44.3369 100.63 46.3198C100.63 48.2667 100.215 49.9702 99.3862 51.4303C98.575 52.8904 97.4213 54.0171 95.9251 54.8103C94.447 55.6034 92.7074 56 90.7065 56ZM90.7065 50.9977C91.6078 50.9977 92.383 50.8535 93.0319 50.565C93.6809 50.2586 94.1766 49.7629 94.5191 49.0779C94.8796 48.3748 95.0599 47.4374 95.0599 46.2657C95.0599 45.094 94.8886 44.1747 94.5461 43.5077C94.2217 42.8227 93.7349 42.336 93.086 42.0475C92.437 41.7411 91.6439 41.5879 90.7065 41.5879C89.8232 41.5879 89.0571 41.7411 88.4081 42.0475C87.7592 42.336 87.2544 42.8227 86.8939 43.5077C86.5514 44.1927 86.3801 45.1301 86.3801 46.3198C86.3801 47.4735 86.5424 48.3928 86.8669 49.0779C87.2094 49.7629 87.7051 50.2586 88.354 50.565C89.003 50.8535 89.7871 50.9977 90.7065 50.9977Z" fill="#FAFAFA"/>
+            <path className='fill-zinc-50' d="M103.322 55.7567V36.8289H112.975C114.399 36.8289 115.625 37.0813 116.652 37.586C117.68 38.0907 118.473 38.8028 119.032 39.7221C119.609 40.6415 119.897 41.7321 119.897 42.9939C119.897 44.1837 119.654 45.1841 119.167 45.9953C118.698 46.7885 118.058 47.4284 117.247 47.9151C116.454 48.3838 115.544 48.7263 114.516 48.9427C113.507 49.1409 112.47 49.2401 111.407 49.2401H108.784V55.7567H103.322ZM114.922 55.7567L111.082 48.8345L116.22 47.8881L120.952 55.7567H114.922ZM108.784 45.4275H111.866C112.768 45.4275 113.435 45.2562 113.867 44.9137C114.3 44.5712 114.516 44.0485 114.516 43.3454C114.516 42.6604 114.3 42.1647 113.867 41.8583C113.453 41.5338 112.786 41.3715 111.866 41.3715H108.784V45.4275Z" fill="#FAFAFA"/>
+            <path className='fill-zinc-50' d="M22.6364 8.31437C17.3604 11.3605 17.3604 18.9758 22.6364 22.0219L27.9436 25.086C27.0779 29.3505 22.178 31.7563 18.1848 29.4508L3.34592 20.8836C-1.05078 18.3451 -1.05077 11.9991 3.34593 9.46062L18.1848 0.893406C22.1758 -1.41079 27.0724 0.991069 27.9421 5.25111L22.6364 8.31437Z" fill="#FAFAFA"/>
+            <path className='fill-zinc-50' d="M42.7513 7.75606C37.0356 11.056 37.0356 19.3059 42.7513 22.6059L47.3679 25.2713C45.8292 29.2048 40.9895 31.207 36.9806 28.8925L24.1203 21.4675C19.2839 18.6753 19.2839 11.6946 24.1203 8.90231L36.9806 1.47739C40.9874 -0.835903 45.8239 1.16286 47.3654 5.0921L42.7513 7.75606Z" fill="#FAFAFA"/>
+            <path className='fill-zinc-50' d="M62.2067 7.19773C56.0513 10.7515 56.0513 19.6361 62.2067 23.1899L65.9807 25.3688C63.9338 28.9933 59.1417 30.6578 55.117 28.3341L44.2352 22.0515C38.9592 19.0054 38.9592 11.3901 44.2352 8.34398L55.117 2.06136C59.1395 -0.260999 63.9286 1.40041 65.9773 5.02077L62.2067 7.19773Z" fill="#FAFAFA"/>
+            <path className='fill-zinc-50' d="M81.0025 6.55697C74.4074 10.3646 74.4074 19.8838 81.0025 23.6914L83.8216 25.319C81.3795 28.6513 76.6303 30.0239 72.5938 27.6934L63.6905 22.5531C57.9748 19.2531 57.9748 11.0032 63.6905 7.70322L72.5938 2.56289C76.6281 0.233707 81.3742 1.6035 83.8175 4.93172L81.0025 6.55697Z" fill="#FAFAFA"/>
+            <path className='fill-zinc-50' d="M99.1389 5.99865C92.1042 10.0601 92.1042 20.2139 99.1389 24.2754L100.865 25.2721C98.1042 28.2848 93.4297 29.3727 89.4112 27.0526L82.4864 23.0546C76.331 19.5008 76.331 10.6163 82.4864 7.06246L89.4112 3.06442C93.4753 0.718021 98.2102 1.85735 100.958 4.94811L99.1389 5.99865Z" fill="#FAFAFA"/>
+            <path className='fill-zinc-50' d="M100.623 23.6386C94.0276 19.8309 94.0276 10.3118 100.623 6.50414L105.569 3.6484C112.164 -0.159254 120.408 4.60032 120.408 12.2156V17.9271C120.408 25.5424 112.164 30.3019 105.569 26.4943L100.623 23.6386Z" fill="#FAFAFA"/>
+        </svg>
+
+        {/* Footer Navigation */}
+        <div className="flex flex-1 items-end md:h-14">
+            <ul className="flex flex-col items-center md:flex-row md:items-start gap-4">
+                {footerNavItems.map(({ href, label }, i) => (
+                    <li key={`footer-nav-item-${i}`} className='inline-flex'>
+                        <Link href={href} className="font-lato font-bold text-xs uppercase leading-none text-zinc-50 hover:underline">{label}</Link>
+                    </li>
+                ))}
+            </ul>
+        </div>
+    </footer>
+  );
+};
