+++ src/components/UI/Header.js
@@ -1,0 +1,43 @@
+'use client';
+
+import { useEffect, useRef, useState } from 'react';
+import Link from 'next/link';
+import HeaderNav from '@/components/UI/HeaderNav';
+
+export default function Header() {
+  const [isSticky, setIsSticky] = useState(false);
+  const topSpacerRef = useRef(null);
+
+  useEffect(() => {
+    const observer = new IntersectionObserver( ([entry]) => { setIsSticky(!entry.isIntersecting); }, { threshold: [0] } );
+    topSpacerRef.current && observer.observe(topSpacerRef.current);
+    return () => { topSpacerRef.current && observer.unobserve(topSpacerRef.current); };
+  }, []);
+
+  return (
+    <>
+      <div ref={topSpacerRef} className={`h-6 transition-colors ${isSticky ? 'bg-zinc-50' : 'bg-transparent'}`}></div> {/* top-spacer */}
+      <header className={`sticky top-0 z-30 flex justify-between items-center h-14 px-4 md:px-8 transition-colors ${isSticky ? 'bg-zinc-50' : 'bg-none'}`}>
+        <Link href="/" className="relative z-40 flex items-center">
+          <svg className={`absolute w-[142.5px] h-[24px] md:w-[190px] md:h-[32px] transition-opacity ${!isSticky ? 'opacity-0' : 'opacity-100 duration-600'}`} viewBox="0 0 190 32" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path className='fill-zinc-800' d="M16.4074 31.2598C12.9503 31.2598 10.0032 30.6364 7.56613 29.3895C5.1291 28.1426 3.25882 26.3715 1.95529 24.0762C0.651764 21.7809 0 19.0888 0 16C0 13.478 0.425063 11.2676 1.27519 9.36901C2.12532 7.4704 3.30133 5.88349 4.80322 4.6083C6.33344 3.33311 8.10454 2.36964 10.1165 1.71787C12.1285 1.06611 14.2963 0.740227 16.62 0.740227C18.0369 0.740227 19.4396 0.881914 20.8281 1.16529C22.245 1.42033 23.6335 1.83122 24.9937 2.39798L26.3539 2.95056L24.9937 9.284L23.251 8.98646C22.0608 8.75976 20.9415 8.58973 19.893 8.47638C18.8445 8.36303 17.8527 8.30636 16.9175 8.30636C15.2173 8.30636 13.7579 8.56139 12.5394 9.07147C11.3492 9.55321 10.4282 10.3608 9.77646 11.4943C9.12469 12.6278 8.79881 14.1581 8.79881 16.085C8.79881 17.927 9.09636 19.4288 9.69144 20.5907C10.2865 21.7242 11.1792 22.546 12.3693 23.0561C13.5595 23.5661 15.0047 23.8212 16.705 23.8212C17.4418 23.8212 18.1785 23.7928 18.9153 23.7362C19.6804 23.6511 20.4455 23.5378 21.2107 23.3961V19.7831H14.3671V13.7897H28.5643V28.0718C27.6008 28.6952 26.4531 29.2478 25.1212 29.7295C23.8177 30.183 22.415 30.5513 20.9131 30.8347C19.4112 31.1181 17.9093 31.2598 16.4074 31.2598Z" fill="#27272A"/>
+            <path className='fill-zinc-800' d="M32.6914 30.8772V1.12278H56.5374V8.47638H41.2776V12.5145H54.8372V19.018H41.2776V23.4811H56.835V30.8772H32.6914Z" fill="#27272A"/>
+            <path className='fill-zinc-800' d="M60.6276 30.8772V1.12278H68.2788L80.4781 16.2975H80.5631V1.12278H89.0644V30.8772H81.7958L69.2564 15.2349H69.1714V30.8772H60.6276Z" fill="#27272A"/>
+            <path className='fill-zinc-800' d="M94.2923 30.8772V1.12278H102.879V11.9619H114.015V1.12278H122.602V30.8772H114.015V19.4855H102.879V30.8772H94.2923Z" fill="#27272A"/>
+            <path className='fill-zinc-800' d="M130.298 31.2598L126.982 27.9868L154.569 0.740227L157.884 4.01322L130.298 31.2598ZM142.455 31.2598C139.224 31.2598 136.433 30.6364 134.081 29.3895C131.757 28.1143 129.972 26.329 128.725 24.0337C127.478 21.7384 126.855 19.0463 126.855 15.9575C126.855 12.8687 127.507 10.1908 128.81 7.9238C130.114 5.62846 131.941 3.85736 134.293 2.61051C136.645 1.36365 139.366 0.740227 142.455 0.740227C145.713 0.740227 148.505 1.37782 150.828 2.65301C153.152 3.89987 154.937 5.67096 156.184 7.9663C157.431 10.2333 158.054 12.9254 158.054 16.0425C158.054 19.103 157.403 21.7809 156.099 24.0762C154.824 26.3715 153.01 28.1426 150.658 29.3895C148.335 30.6364 145.6 31.2598 142.455 31.2598ZM142.455 23.3961C143.872 23.3961 145.09 23.1694 146.11 22.716C147.13 22.2343 147.91 21.455 148.448 20.3782C149.015 19.273 149.298 17.7994 149.298 15.9575C149.298 14.1156 149.029 12.6703 148.491 11.6219C147.981 10.545 147.215 9.77991 146.195 9.32651C145.175 8.84477 143.928 8.6039 142.455 8.6039C141.066 8.6039 139.862 8.84477 138.842 9.32651C137.821 9.77991 137.028 10.545 136.461 11.6219C135.923 12.6987 135.654 14.1722 135.654 16.0425C135.654 17.8561 135.909 19.3013 136.419 20.3782C136.957 21.455 137.736 22.2343 138.757 22.716C139.777 23.1694 141.009 23.3961 142.455 23.3961Z" fill="#27272A"/>
+            <path className='fill-zinc-800' d="M162.286 30.8772V1.12278H177.461C179.699 1.12278 181.626 1.51951 183.241 2.31296C184.857 3.10641 186.104 4.22575 186.982 5.67096C187.889 7.11618 188.342 8.8306 188.342 10.8142C188.342 12.6845 187.96 14.2572 187.195 15.5324C186.458 16.7793 185.452 17.7853 184.177 18.5504C182.93 19.2872 181.499 19.8256 179.883 20.1656C178.297 20.4773 176.667 20.6332 174.995 20.6332H170.872V30.8772H162.286ZM180.521 30.8772L174.485 19.9956L182.561 18.5079L190 30.8772H180.521ZM170.872 14.6398H175.718C177.135 14.6398 178.183 14.3706 178.863 13.8322C179.543 13.2938 179.883 12.472 179.883 11.3668C179.883 10.29 179.543 9.5107 178.863 9.02896C178.212 8.51889 177.163 8.26385 175.718 8.26385H170.872V14.6398Z" fill="#27272A"/>
+          </svg>
+          <svg className={`w-[142.5px] h-[36px] md:w-[190px] md:h-[48px] transition-opacity duration-300 ${isSticky ? 'opacity-0' : 'opacity-100'}`} viewBox="0 0 190 48" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path className={`fill-zinc-800 transition-opacity ${!isSticky ? 'opacity-100 duration-75 delay-[450ms]' : 'opacity-0'}`} d="M35.6575 13.1521C27.3288 17.9706 27.3288 30.0169 35.6575 34.8354L44.0354 39.6823C42.6688 46.4281 34.9338 50.2337 28.6301 46.5868L5.20547 33.0347C-1.73517 29.0193 -1.73515 18.9807 5.20549 14.9653L28.6302 1.41324C34.9304 -2.23167 42.6603 1.56772 44.0332 8.30648L35.6575 13.1521Z" fill="#27272A"/>
+            <path className={`fill-zinc-800 transition-opacity ${!isSticky ? 'opacity-100 duration-75 delay-[375ms]' : 'opacity-0'}`} d="M67.4109 12.2689C58.3881 17.489 58.3881 30.5391 67.411 35.7591L74.6988 39.9754C72.2698 46.1977 64.6299 49.3649 58.3014 45.7036L38 33.9584C30.3653 29.5415 30.3653 18.4991 38 14.0821L58.3014 2.33699C64.6264 -1.32229 72.2615 1.83946 74.6948 8.05493L67.4109 12.2689Z" fill="#27272A"/>
+            <path className={`fill-zinc-800 transition-opacity ${!isSticky ? 'opacity-100 duration-75 delay-[300ms]' : 'opacity-0'}`} d="M98.1233 11.3857C88.4064 17.0073 88.4064 31.0613 98.1233 36.6829L104.081 40.1296C100.85 45.863 93.2848 48.496 86.9315 44.8204L69.7534 34.8822C61.4247 30.0637 61.4247 18.0174 69.7534 13.1989L86.9315 3.26074C93.2814 -0.412883 100.841 2.21522 104.076 7.94209L98.1233 11.3857Z" fill="#27272A"/>
+            <path className={`fill-zinc-800 transition-opacity ${!isSticky ? 'opacity-100 duration-75 delay-[225ms]' : 'opacity-0'}`} d="M127.795 10.3721C117.384 16.3953 117.384 31.4531 127.795 37.4762L132.245 40.0509C128.39 45.322 120.893 47.4933 114.521 43.8068L100.466 35.6756C91.4429 30.4555 91.4429 17.4054 100.466 12.1853L114.521 4.05409C120.889 0.369665 128.381 2.53647 132.238 7.80122L127.795 10.3721Z" fill="#27272A"/>
+            <path className={`fill-zinc-800 transition-opacity ${!isSticky ? 'opacity-100 duration-75 delay-[150ms]' : 'opacity-0'}`} d="M156.425 9.48894C145.32 15.9136 145.32 31.9753 156.425 38.4L159.15 39.9767C154.791 44.7423 147.412 46.4633 141.068 42.7932L130.137 36.4689C120.42 30.8473 120.42 16.7933 130.137 11.1717L141.069 4.84743C147.484 1.13577 154.959 2.93803 159.297 7.82714L156.425 9.48894Z" fill="#27272A"/>
+            <path className={`fill-zinc-800 transition-opacity ${!isSticky ? 'opacity-100 duration-75 delay-[75ms]' : 'opacity-0'}`} d="M158.767 37.3926C148.356 31.3695 148.356 16.3117 158.767 10.2885L166.575 5.77119C176.986 -0.251949 190 7.27698 190 19.3233V28.358C190 40.4042 176.986 47.9331 166.575 41.91L158.767 37.3926Z" fill="{``}#27272A"/>
+          </svg>
+        </Link>
+        <HeaderNav />
+      </header>
+    </>
+  );
+};
